Initial Setup and Login Tests Creation
Added method to go to the Bluesky URL and click the Sign in button
Added test to enter in just an email and press Next button. Checks for correct error message.
Added test to enter just a password and press Next button. Checks for correct error message.
Added test to enter a bad username and password then click the Next button. Checks for correct error message.
Added test to click the forgot password button. Checks for a specific text message on the page.
Added test to login successfully. Checks that the web URL goes to bsky main page. This needs to be updated to check that the user is actually logged in. Can be done by looking for the profile/going to the profile.
Added Allure reports.